Fix fstat type parameter
Summary:
GCC11 would complain:

newlib/libgloss/kvx-elf/fstat.c: In function ‘__gloss_fstat’:
newlib/libgloss/kvx-elf/fstat.c:44:22: warning: passing argument 2 of ‘sc_fstat’ from incompatible pointer type [-Wincompatible-pointer-types]
   44 |   ret = sc_fstat(fd, res);
      |                      ^~~
      |                      |
      |                      long int *
newlib/libgloss/kvx-elf/fstat.c:37:27: note: expected ‘int *’ but argument is of type ‘long int *’
   37 | int sc_fstat(int fd, int *res);
      |                      ~~~~~^~~

ref T16832

Test Plan: works®

Reviewers: O26 newlib, lthevenoux

Reviewed By: O26 newlib, lthevenoux

Subscribers: lthevenoux

Maniphest Tasks: T16832

Differential Revision: https://phab.kalray.eu/D14157

+++ libgloss/kvx-cos/fstat.c
@@ -6,7 +6,7 @@
 /***************************************************************
  *	fstat
  ***************************************************************/
-int sc_fstat(int fd, int *res);
+int sc_fstat(int fd, uint64_t *res);
 
 static inline int __gloss_fstat(int fd, struct stat *st)
 {
